WIP: automate the macOS-Mojave installation
Use qemu in VNC server mode and then use vncdotool to drive the
installation.

The goal is to have a booted, snapshotted VM that can easily be resumed
to get fast startup times. qemu takes ~1s to boot a snapshotted Ubuntu
VMs. If that also works for macOS is means that we can spawn on-demand
VMs quickly and have a fresh install to run automated tests for example.

Usage:

start ./install-macOS-NG.sh in one terminal
start ./automation.sh in another
follow the instructions

once installed, start ./boot-macOS-NG.sh to start the VM without the
install media.

TODO:
* actually detect the screens, vncdotools's expect seems to be broken
* actually create a resumable snapshot of the booted VM
* for some reason the installer now fails on reboot. See cap8-error.png
* currently the output image is fairly large (20GB). Is there any way to
  reduce that to a reasonable size?

At some point I want to automate all of the installation process with
Nix as the driver for the build.