Add OpenCore-Catalina image configured with ShowPicker=false
The current OpenCore image always displays the disk selection menu, as
`ShowPicker` is `true`, and doesn't provide a timeout, making it
impossible to boot to macOS on a headless environment (i.e. CI).

This commit creates a new OpenCore `config.plist` that has `ShowPicker`
set to `false`, and includes the resulting `qcow2` image.

Signed-off-by: Juan Cruz Viotti <jv@jviotti.com>